requirements.txt
Django>=1.11
requests
python-decouple
Sphinx>=1.8.1
sphinx-rtd-theme==<0 class="1 8"></0>
tox
